requirements.txt
asgiref==3.5.0
attrs==23.1.0
bcrypt==3.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cloudinary==1.32.0
dj-database-url==0.5.0
Django==3.2.16
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.4
jsonschema==4.6.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.19.3
python-decouple==3.5
python-dotenv==0.19.2
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.8
whitenoise==6.2.0
